+++ apps/web/app/components/blocks/UtilityBar/partials/UtilityBarSectionEditor.vue
