+++ apps/web/app/components/blocks/UtilityBar/partials/UtilityBarSectionsList.vue
